refactor: update class name of qwen bot
