+++ bot/bot_factory.py
@@ -45,6 +45,6 @@ def create_bot(bot_type):
         return ClaudeAIBot()
 
     elif bot_type == const.QWEN:
-        from bot.tongyi.tongyi_qwen_bot import TongyiQwenBot
-        return TongyiQwenBot()
+        from bot.ali.ali_qwen_bot import AliQwenBot
+        return AliQwenBot()
     raise RuntimeError
